+++ src/renderer/rendererutils.ts
@@ -704,7 +704,28 @@ const getFriendlyEncoderName = (enc: ESupportedEncoders) => {
     case ESupportedEncoders.VAAPI_AV1:
       return 'VAAPI AV1';
     default:
-      throw new Error('Unknown Encoder');
+      throw new Error('Unknown Encoder: ' + enc);
+  }
+};
+
+const getFriendlyCodecName = (enc: string) => {
+  switch (enc) {
+    case ESupportedEncoders.OBS_X264:
+    case ESupportedEncoders.NVENC_H264:
+    case ESupportedEncoders.AMD_H264:
+    case ESupportedEncoders.QSV_H264:
+    case ESupportedEncoders.VAAPI_H264:
+      return 'H264';
+    case ESupportedEncoders.NVENC_H265:
+    case ESupportedEncoders.AMD_H265:
+      return 'H265';
+    case ESupportedEncoders.NVENC_AV1:
+    case ESupportedEncoders.AMD_AV1:
+    case ESupportedEncoders.QSV_AV1:
+    case ESupportedEncoders.VAAPI_AV1:
+      return 'AV1';
+    default:
+      throw new Error('Unknown Encoder2: ' + enc);
   }
 };
 
@@ -1209,4 +1230,5 @@ export {
   videoMatch,
   videoMatchName,
   translateQuality,
+  getFriendlyCodecName,
 };